+++ src/App.tsx
@@ -1,8 +1,10 @@
 import React from 'react';
 import { BrowserRouter, Routes, Route, Navigate } from 'react-router-dom';
-import Home from '@src/pages/home';
-import Detail from '@src/pages/detail';
-import Mypage from '@src/pages/myPage';
+
+import Detail from './pages/Detail';
+import Home from './pages/Home';
+import Mypage from './pages/Mypage';
+
 
 const App: React.FC = () => {
   return (
