+++ src/components/modal/index.tsx
@@ -1,5 +1,5 @@
-import React, { useEffect, useState } from 'react';
 import { Modal } from 'antd';
+import React, { useEffect, useState } from 'react';
 import '@assets/modal.scss';
 
 interface ModalProps {
