+++ src/components/search/index.tsx
@@ -1,6 +1,6 @@
-import React, { Dispatch, useState, SetStateAction } from 'react';
 import { Select } from 'antd';
 import { AxiosResponse } from 'axios';
+import React, { Dispatch, useState, SetStateAction } from 'react';
 
 const { Option } = Select;
