+++ src/pages/Home/matchSelector/dateSelect.tsx
@@ -1,10 +1,11 @@
-import React, { useContext, useState, useEffect } from 'react';
-import { DatePicker as AntdDatePicker, ConfigProvider } from 'antd';
-import type { DatePickerProps } from 'antd';
-import 'moment/locale/ko';
+import { DatePicker as AntdDatePicker, ConfigProvider, DatePickerProps } from 'antd';
 import locale from 'antd/es/locale/ko_KR';
-import { MatchInfoContext } from './MatchInfoProvider';
 import moment from 'moment';
+import React, { useContext, useState, useEffect } from 'react';
+
+import { MatchInfoContext } from './MatchInfoProvider';
+
+import 'moment/locale/ko';
 
 const DatePicker = () => {
   const { matchData, setMatchData } = useContext(MatchInfoContext);
